Disovery Enhancements - backed #87

+++ log_manager/views.py
@@ -1,19 +1,21 @@
 import ast
 import datetime
 import json
-import traceback
 
-from celery.result import AsyncResult
+from celery import states
 from django.core.paginator import Paginator, EmptyPage
 from django_celery_results.models import TaskResult
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.request import Request
 from rest_framework.response import Response
+
+from log_manager.logger import get_backend_logger
 from log_manager.models import Logs
+from orca_backend.celery import cancel_task
 
+_logger = get_backend_logger()
 
-# Create your views here.
 
 @api_view(['get'])
 def get_logs(request: Request, **kwargs):
@@ -45,11 +47,10 @@ def get_logs(request: Request, **kwargs):
         )
         return Response(final_result, status=status.HTTP_200_OK)
     except EmptyPage as e:
-        print(str(e))
+        _logger.error("EmptyPage Error: ", e)
         return Response({"message": str(e)}, status=status.HTTP_204_NO_CONTENT)
     except Exception as e:
-        print(str(e))
-        print(traceback.format_exc())
+        _logger.error("Error: ", e)
         return Response({"message": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
@@ -66,11 +67,24 @@ def delete_logs(request: Request, **kwargs):
     - If fails returns a JSON response with 500 status.
     """
     try:
-        Logs.objects.all().delete()
+        data = request.data
+        log_ids = data.get("log_ids", [])
+        task_ids = data.get("task_ids", [])
+
+        if log_ids:
+            _logger.debug("Deleting logs for given ids: %s", log_ids)
+            Logs.objects.filter(id__in=log_ids).delete()
+        if task_ids:
+            _logger.debug("Deleting celery tasks: %s", task_ids)
+            delete_celery_tasks_data(task_ids)
+
+        if not log_ids and not task_ids:
+            _logger.info("Deleting all logs")
+            Logs.objects.all().delete()
+            delete_celery_tasks_data()
         return Response({"message": "deleted successfully."}, status=status.HTTP_200_OK)
     except Exception as e:
-        print(str(e))
-        print(traceback.format_exc())
+        _logger.error("Error: ", e)
         return Response({"message": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
@@ -84,7 +98,10 @@ def get_celery_tasks_data() -> list:
     task_results = TaskResult.objects.all()
     result_data = []
     for result in task_results:
-        task_kwargs = ast.literal_eval(result.task_kwargs.strip('\"')) if result.task_kwargs else {}
+        try:
+            task_kwargs = ast.literal_eval(result.task_kwargs.strip('\"')) if result.task_kwargs else {}
+        except ValueError:
+            task_kwargs = {"result": result.task_kwargs}
         http_path = task_kwargs.pop("http_path", "")
         result_data.append(
             {
@@ -100,3 +117,24 @@ def get_celery_tasks_data() -> list:
             }
         )
     return result_data
+
+
+def delete_celery_tasks_data(task_ids: list = None) -> None:
+    """
+    function to delete celery tasks data from database
+
+    Parameters:
+        - task_ids: list of task ids
+    Returns:
+        - None
+    """
+    tasks = TaskResult.objects.exclude(status=states.STARTED)
+    if task_ids:
+        tasks = tasks.filter(task_id__in=task_ids)
+    for i in tasks:
+        # cancel task if not started
+        if i.status == states.PENDING:
+            cancel_task(i.task_id)
+
+    # delete all task
+    tasks.delete()
